warning message while compiling has removed
git-svn-id: svn+ssh://source.eecs.berkeley.edu/home/svn/chess/pret/trunk/pret_iss@1541 2fe1bebc-546c-4154-8a90-0a651ce2c4e0

+++ soft/graphic/thread1.c
@@ -126,7 +126,7 @@ INT main( )
 				DEADPLLI( "10" );
 				IOWR_32DIRECT( IO_ADDR_VGA, 0, *p_color_line );
 #endif
-			} while( p_color_line++ < limit );
+			} while( ( int ) p_color_line++ < ( int ) limit );
 
 			// process front-porch pixel
 			DEADPLL( N_VGA_CLK_H_FP );
